+++ py_slides_term/endtoend/configs/analysis.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+from .base import BaseLayerConfig
+
+
+@dataclass(frozen=True)
+class AnalysisLayerConfig(BaseLayerConfig):
+    use_cache: bool = True
